+++ docs/make.jl
@@ -16,7 +16,8 @@ makedocs(;
         assets = String[]
     ),
     pages = [
-        "Home" => "index.md"
+        "Home" => "index.md",
+        "API" => "api.md"
     ]
 )